sultan-ai-v12.7-live/requirements.txt
streamlit==1.32.0
plotly==5.22.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
textblob>=0.17.0
vaderSentiment>=3.3.2
python-dotenv>=1.0.0
scikit-learn>=1.3.0
